+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- **`vtex.styleguide.d.ts`**
+  - Typing for styleguide.
 
 ## [2.3.1] - 2018-10-29
 ### Added