+++ CHANGELOG.md
@@ -6,9 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
+### Removed
 - **`pages.json`**
-  - Removed redirects
+  - Redirects on sidebar menu.
 
 ## [2.3.0] - 2018-10-24
 ### Changed